pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
